fix(ronda 27): precision milimetrica en pagos (monto decimal end-to-end)
- cliente enviaba price.toInt(): SAGEN PASS (9.90) llegaba como 9 al
  server y este rechazaba con 400 (pkg.amount !== amount), y cualquier
  donacion con decimales se truncaba
- createPreference/registerPendingPayment pasan a double y usan price
- webhook: parseInt -> parseFloat ('9.9' se acreditaba como 9,0)
- earned-totales donatedBefore/After: toInt -> round (sin perder centimos)
- descripcion del item: $\ -> S/ con 2 decimales (PEN)
- test jest: SAGEN PASS decimal acredita 9.90, no 9

+++ lib/services/mercado_pago_service.dart
@@ -28,7 +28,7 @@ class MercadoPagoService {
       AppConfig.mercadopagoFunctionsUrl.replaceAll(RegExp(r'/+$'), '');
 
   Future<MercadoPagoPreference> createPreference({
-    required int amount,
+    required double amount,
     required String productId,
     required String idToken,
     double? price,
@@ -109,7 +109,7 @@ class MercadoPagoService {
     required String paymentMethod,
     required String operationId,
     required String idToken,
-    int? amount,
+    double? amount,
     String? productId,
   }) async {
     final url = Uri.parse('$_baseUrl/api/registerPendingPayment');
